pyproject.toml
[project]
name = "flow-atelier"
version = "0.1.0"
description = "CLI tool and async workflow engine for running reproducible DAG workflows (conduits)."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "typer>=0.12",
    "pyyaml>=6.0",
    "pydantic>=2.8",
    "pydantic-settings>=2.4",
    "rich>=13.7",
    "agent-client-protocol>=0.9.0",
]

[project.scripts]
atelier = "app.main:app"

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-timeout>=2.3",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
timeout = 300

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
